Rubric questions: allow weight for each cell: Remove legacy code #9026 (#9028)
* Remove legacy code and test

* Refactor test data according to new format

* Convert remaining rubric test data to new format

* Add migration script to remove 1D weight list from questionMetaData

* Delete DataMigrationForRubricQuestionsWithWeightsAttached file
